Move CmisManagerTask into it's own module
