+++ sonic-xcvrd/xcvrd/cmis/__init__.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+
+"""
+    cmis
+    CMIS transceiver management module for SONiC
+"""
+
+from .cmis_manager_task import CmisManagerTask
+
+__all__ = ['CmisManagerTask']
