+++ sonic-xcvrd/xcvrd/xcvrd_utilities/common.py
@@ -19,6 +19,25 @@ except ImportError as e:
 
 SYSLOG_IDENTIFIER_COMMON = "common"
 
+# CMIS States
+CMIS_STATE_UNKNOWN = 'UNKNOWN'
+CMIS_STATE_INSERTED = 'INSERTED'
+CMIS_STATE_DP_PRE_INIT_CHECK = 'DP_PRE_INIT_CHECK'
+CMIS_STATE_DP_DEINIT = 'DP_DEINIT'
+CMIS_STATE_AP_CONF = 'AP_CONFIGURED'
+CMIS_STATE_DP_ACTIVATE = 'DP_ACTIVATION'
+CMIS_STATE_DP_INIT = 'DP_INIT'
+CMIS_STATE_DP_TXON = 'DP_TXON'
+CMIS_STATE_READY = 'READY'
+CMIS_STATE_REMOVED = 'REMOVED'
+CMIS_STATE_FAILED = 'FAILED'
+
+CMIS_TERMINAL_STATES = {
+    CMIS_STATE_FAILED,
+    CMIS_STATE_READY,
+    CMIS_STATE_REMOVED
+}
+
 # Global variables that will be injected from the parent module
 platform_chassis = None
 platform_sfputil = None
